make potassium_sorbate an instance of classes ['Thing', 'Sauce'] from AI answer
Query: most adequate class for 'Potassium sorbate' among: 'Food Dressing', 'Food Seasoning', 'Sauce', 'Spice', 'Sweetener'. concise.
Answer:
	The most adequate class for "Potassium sorbate" among the listed options is "Sauce".

Potassium sorbate is commonly used as a preservative in sauces. It is an anti-microbial agent that helps to prevent the growth of bacteria in food.
meaning that the most adequate type for potassium_sorbate is Sauce.
Current classes of potassium_sorbate are:
- Thing
- Sauce
Query cache in files:
- /kgfiller/data/cache-e2861d4b05a015798a5572e71364d591249446cf91445c25dafe265760e078d2.yml