make yogurt an instance of classes ['Thing', 'Milk'] from AI answer
Query: most adequate class for 'Yogurt' among: 'Butter', 'Cheese', 'Milk', 'Yogurt'. concise.
Answer:
	Sure, here is my answer:

**Yogurt** is the most adequate class for "Yogurt" among "Butter," "Cheese," "Milk," and "Yogurt."

Yogurt is a fermented dairy product, made from cow's milk and other ingredients. It is a good source of protein, calcium, and probiotics. Yogurt can be eaten plain or used in a variety of recipes.
meaning that the most adequate type for yogurt is Milk.
Current classes of yogurt are:
- Thing
- Milk
Query cache in files:
- /kgfiller/data/cache-ab08ff8cbff6a710ece45ee121bef338135459cbc2f6df1f17e427bbadb366ad.yml