make lobster an instance of classes ['Thing', 'Fish'] from AI answer
Query: most adequate class for 'Lobster' among: 'Edible Crustaceans', 'Edible Fish', 'Edible Molluscs'. concise.
Answer:
	Sure, here is the answer:

The most adequate class for "Lobster" among "Edible Crustaceans," "Edible Fish," and "Edible Molluscs" is "Edible Crustaceans."

Lobster is a type of crustacean, which is a group of arthropods that includes shrimp, crabs, and crayfish. Lobsters are often considered to be one of the most flavorful and healthy seafood species. They are rich in protein, low in fat, and contain a variety of vitamins and minerals.
meaning that the most adequate type for lobster is Fish.
Current classes of lobster are:
- Thing
- Fish
Query cache in files:
- /kgfiller/data/cache-9dc1f44bfaa9cbaea104e4340709d84e3a0b807bc1d8798a6e146dbb05a69863.yml