make lemon_ginger an instance of classes ['Thing', 'Tea'] from AI answer
Query: most adequate class for 'Lemon Ginger' among: 'Coffee', 'Tea'. concise.
Answer:
	**Answer:**

The most adequate class for "Lemon Ginger" among "Coffee" and "Tea" is Tea.

Lemon Ginger is a soothing and digestive beverage that is commonly prepared with tea as the base.
meaning that the most adequate type for lemon_ginger is Tea.
Current classes of lemon_ginger are:
- Thing
- Tea
Query cache in files:
- /kgfiller/data/cache-3c29c2b53f5ed4e6913a65ac38e17fb1834156d859a3750898eda17e4974a6c7.yml