merged mayo and mayonnaise together
Query: in the Food Dressing class, should the instances 'Mayo' and 'Mayonnaise' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Mayo and mayonnaise are semantic and ontologic duplicates. They are the same food condiment.
meaning that instances mayo and mayonnaise are semantically identical.
Query cache in files:
- /kgfiller/data/cache-90b4aa0a97f73a561eff5238ef3da007546bf9dbd23bd49aded9caa3ff77f7ad.yml