merged pepper and wrapper together
Query: in the Food Dressing class, should the instances 'Pepper' and 'Wrapper' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

The words "Pepper" and "Wrapper" are semantically and ontologically duplicates. They are both words used to describe physical substances that are added to food to impart flavor or protect it from spoilage, respectively.
meaning that instances pepper and wrapper are semantically identical.
Query cache in files:
- /kgfiller/data/cache-5a893f9c36484b8c0344d7128a4eab59bf669cb88f2fb55029020356dde6d594.yml