instances instances_of_the_sauce_class and red_sauce were NOT merged together
Query: in the Sauce class, should the instances 'Instances of the Sauce class' and 'Red sauce' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

'Instances of the Sauce class' and 'Red sauce' are not semantic and ontologic duplicates because they represent different concepts.

* **Instances of the Sauce class** is a group of objects that are instances of the Sauce class. This concept is related to the class definition and its instantiation.
* **Red sauce** is a specific type of sauce, which is one of many possible types of sauce. This concept is related to the domain of sauces and their various classifications.

Therefore, although they are both related to the concept of sauce, they represent different ontological categories.
meaning that instances instances_of_the_sauce_class and red_sauce are different.
Query cache in files:
- /kgfiller/data/cache-24ad8284c4ec1a8e6e1ee199cf51b03df3c116136d24c764b01af7c4603198ec.yml