merged instances_of_the_sauce_class and sawa_sauce together
Query: in the Sauce class, should the instances 'Instances of the Sauce class' and 'Sawa sauce' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes

The instances 'Instances of the Sauce class' and 'Sawa sauce' are semantic and ontologic duplicates. They are essentially the same concept, referring to the same sauce.
meaning that instances instances_of_the_sauce_class and sawa_sauce are semantically identical.
Query cache in files:
- /kgfiller/data/cache-fa14d1be762fc1ba8894fc1f186a7ceb41cc378cc9269126303d153432e57859.yml