merged instances_of_the_sauce_class and fish_sauce together
Query: in the Sauce class, should the instances 'Fish sauce' and 'Instances of the Sauce class' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Fish sauce and "Instances of the Sauce class" are semantic and ontologic duplicates as they both refer to the same concept, which is fish sauce.
meaning that instances instances_of_the_sauce_class and fish_sauce are semantically identical.
Query cache in files:
- /kgfiller/data/cache-3746dc5ef0dd9dabb560e6b36d3c39abcefeb7603c89cb58553632cda59d72ba.yml