merged instances_of_the_sauce_class and tomato_sauce together
Query: in the Sauce class, should the instances 'Instances of the Sauce class' and 'Tomato sauce' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

'Instances of the Sauce class' and 'Tomato sauce' are semantic and ontologic duplicates, as they represent the same concept, namely tomato sauce.
meaning that instances instances_of_the_sauce_class and tomato_sauce are semantically identical.
Query cache in files:
- /kgfiller/data/cache-9b53919029ab04c174e59bdfe290fb05e1d530aa5abd5b6658568f970477a0e7.yml