instances instances_of_the_sauce_class and hot_sauce were NOT merged together
Query: in the Sauce class, should the instances 'Hot Sauce' and 'Instances of the Sauce class' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

'Hot Sauce' and 'Instances of the Sauce class' are not semantic and ontologic duplicates. They represent different concepts:

* **Hot Sauce**: A specific type of sauce.
* **Instances of the Sauce class**: A group of objects that conform to the Sauce class definition.

While they may be related, they are not duplicates.
meaning that instances instances_of_the_sauce_class and hot_sauce are different.
Query cache in files:
- /kgfiller/data/cache-2e73da9482a02206c1b8744e600332e53c8f5dd6236532f307a1d3f5ece7c13f.yml