instances instances_of_the_sauce_class and teriyaki_sauce were NOT merged together
Query: in the Sauce class, should the instances 'Instances of the Sauce class' and 'Teriyaki sauce' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

"Instances of the Sauce class" and "Teriyaki sauce" are not semantic and ontologic duplicates. They are different entities.

* **Instances of the Sauce class** is a group of objects that belong to the Sauce class. This group includes various sauces, such as teriyaki sauce, tomato sauce, and barbecue sauce.
* **Teriyaki sauce** is a specific type of sauce that belongs to the Sauce class. It is a Japanese sauce made with soy sauce, sugar, mirin, and sake.
meaning that instances instances_of_the_sauce_class and teriyaki_sauce are different.
Query cache in files:
- /kgfiller/data/cache-398fa6761d48bec9e81187c7584788dc40d5750f1b210e685927f90757e4fd9d.yml