merged instances_of_the_sauce_class and hoisin_sauce together
Query: in the Sauce class, should the instances 'Hoisin sauce' and 'Instances of the Sauce class' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Hoisin sauce is an instance of the Sauce class. So, "Hoisin sauce" and "Instances of the Sauce class" are semantic and ontologic duplicates.
meaning that instances instances_of_the_sauce_class and hoisin_sauce are semantically identical.
Query cache in files:
- /kgfiller/data/cache-d29789241d8f512e1f4e0f14fbb814f13f8dcbf391407ab14684977ea91c7d73.yml