instances instances_of_the_sauce_class and gluten_free_soy_sauce were NOT merged together
Query: in the Sauce class, should the instances 'Gluten-free soy sauce' and 'Instances of the Sauce class' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

The instances 'Gluten-free soy sauce' and 'Instances of the Sauce class' are not semantic and ontologic duplicates. They are different instances of the Sauce class.
meaning that instances instances_of_the_sauce_class and gluten_free_soy_sauce are different.
Query cache in files:
- /kgfiller/data/cache-b3e2208131c4110b1afc22bd87a627a9db72cf78f5f908aee2a120d5b4ad320c.yml