merged potassium_sorbate and sodium_bicarbonate together
Query: in the Sauce class, should the instances 'Potassium sorbate' and 'Sodium bicarbonate' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Potassium sorbate and sodium bicarbonate are semantic and ontologic duplicates as they are both ingredients used in sauces.
meaning that instances potassium_sorbate and sodium_bicarbonate are semantically identical.
Query cache in files:
- /kgfiller/data/cache-79b2eaef665c17303b3b0cdb77ded96ca0706d790d6436d7dbffc9a75d619c0f.yml