instances sorbic_acid and phosphoric_acid were NOT merged together
Query: in the Sauce class, should the instances 'Phosphoric acid' and 'Sorbic acid' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Phosphoric acid and Sorbic acid are not semantic and ontologic duplicates. They are two different chemicals with distinct chemical structures, properties, and functions in the body.
meaning that instances sorbic_acid and phosphoric_acid are different.
Query cache in files:
- /kgfiller/data/cache-e21ac98a5a05c26ac68a1a2294aab32a59c0d09061fe696dfb727a7dbfe698e7.yml