instances sawa_sauce and worcestershire_sauce were NOT merged together
Query: in the Sauce class, should the instances 'Sawa sauce' and 'Worcestershire sauce' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Sawa sauce and Worcestershire sauce are both semantic and ontologic duplicates. They are both sauces that are used in similar ways in many recipes and are often used in similar combinations. They are also both sauces that are commonly used in British cuisine.
meaning that instances sawa_sauce and worcestershire_sauce are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-b842e3fa6d776999a7de668c1caecf548806552d9b19a92d788d366e6014ad7b.yml