instances worcestershire_sauce and hoisin_sauce were NOT merged together
Query: in the Sauce class, should the instances 'Hoisin sauce' and 'Worcestershire sauce' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Hoisin sauce and Worcestershire sauce are both condiments that are in the same category as other sauces. They are both sweet and savory sauces that are commonly used in Asian cuisine.
meaning that instances worcestershire_sauce and hoisin_sauce are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-83c885b88fd650c9bae5f3b28f2dcfe5f2d544d4f69c86e110cbc6981d61dd11.yml