merged seasoning_cubes and taco_seasoning together
Query: in the Food Seasoning class, should the instances 'Seasoning cubes' and 'Taco seasoning' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

The instances 'Seasoning cubes' and 'Taco seasoning' are semantic and ontologic duplicates. They refer to the same concept, which is seasoning for tacos.
meaning that instances seasoning_cubes and taco_seasoning are semantically identical.
Query cache in files:
- /kgfiller/data/cache-dfe2db94d6bcb929171820a71ede23506a43480cf4c4d35b8c8591974cd97a19.yml