instances seasoning and taco_seasoning were NOT merged together
Query: in the Food Seasoning class, should the instances 'Seasoning' and 'Taco seasoning' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

The instances 'Seasoning' and 'Taco seasoning' are semantic and ontologic duplicates. They refer to the same concept, which is seasoning.
meaning that instances seasoning and taco_seasoning are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-bc40d6ae5287c60c9544a81155dc6ec7bafec5da4d6ff09511d51427468993bd.yml