instances seasoning and egg_roll_seasoning were NOT merged together
Query: in the Food Seasoning class, should the instances 'Egg roll seasoning' and 'Seasoning' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes

The instances 'Egg roll seasoning' and 'Seasoning' are semantic and ontologic duplicates. They are referring to the same concept but different instances of the same concept.
meaning that instances seasoning and egg_roll_seasoning are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-ae7986c11d37326b3d7cbd680c6f3d443296bf2aa06b495aef51e1f4451b3cf5.yml