merged pepper and dried_pepper together
Query: in the Spice class, should the instances 'Dried pepper' and 'Pepper' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Dried pepper and pepper are semantic and ontologic duplicates. They refer to the same food item.
meaning that instances pepper and dried_pepper are semantically identical.
Query cache in files:
- /kgfiller/data/cache-9653593f4bf5809d9be52c55f2265c08d1524fafe85e1bed29d512bf4c3b4f0d.yml