merged allspice and jamaican_allspice together
Query: in the Spice class, should the instances 'Allspice' and 'Jamaican allspice' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Allspice and Jamaican allspice are semantic and ontologic duplicates. They refer to the same spice.
meaning that instances allspice and jamaican_allspice are semantically identical.
Query cache in files:
- /kgfiller/data/cache-eaca3c132593ee09e8e5da2452a79e8a986f02b55794282f1284065159f2983e.yml