instances rice_syrup and corn_syrup were NOT merged together
Query: in the Sweetener class, should the instances 'Corn syrup' and 'Rice Syrup' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Corn syrup and rice syrup are both sweeteners and are ontologically similar. They are both refined starches that are broken down into glucose, which is a simple sugar that is absorbed into the bloodstream.
meaning that instances rice_syrup and corn_syrup are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-b2c08a978f4ab38b97693117563a665d9d22791f5fd36146a80b0d4f8fe4ff78.yml