merged ground_beef and beef_broth together
Query: in the Beef class, should the instances 'Beef broth' and 'Ground Beef' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

The instances 'Beef broth' and 'Ground Beef' are semantic and ontologic duplicates. They are both referring to the same food item, which is beef.
meaning that instances ground_beef and beef_broth are semantically identical.
Query cache in files:
- /kgfiller/data/cache-931c2affdcc013914ff5c8c4c414dbca52a5569b7d7be570df874e8ddd62b11f.yml