instances broth and beef_broth were NOT merged together
Query: in the Beef class, should the instances 'Beef broth' and 'Broth' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Beef broth and broth are semantic and ontologic duplicates. They refer to the same concept, which is beef broth.
meaning that instances broth and beef_broth are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-dcaa6a67c756cfe89cdcd80db1f8c8e201447735c2e5c996a4aad3bcc0099cc6.yml