merged chicken_breast and chicken_wings together
Query: in the Poultry-derived Food class, should the instances 'Chicken Breast' and 'Chicken Wings' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

The instances 'Chicken Breast' and 'Chicken Wings' are semantic and ontologic duplicates because they are both parts of the same whole chicken.
meaning that instances chicken_breast and chicken_wings are semantically identical.
Query cache in files:
- /kgfiller/data/cache-881df7f796f2bb3ada505671b663316eff56e364fb7d21cbeff35818f489714f.yml