merged chicken_breast and chicken together
Query: in the Poultry-derived Food class, should the instances 'Chicken Breast' and 'Chicken' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Chicken breast and chicken are ontological and semantic duplicates, as they refer to the same food source.
meaning that instances chicken_breast and chicken are semantically identical.
Query cache in files:
- /kgfiller/data/cache-8e983a89cca1ce94c39f8e3ea1f5dea5caa11e3993c4e91d9ee890ec5d40d39c.yml