instances chicken_breast and chicken_thighs were NOT merged together
Query: in the Poultry-derived Food class, should the instances 'Chicken Breast' and 'Chicken thighs' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Chicken breast and chicken thighs are not semantic and ontologic duplicates. They are different anatomical parts of the same chicken carcass.
meaning that instances chicken_breast and chicken_thighs are different.
Query cache in files:
- /kgfiller/data/cache-2d0891e38f5849ab9fd57fee94a7cde5e68771a8227764f6268857c96a69dab8.yml